error instead of deprecation
